feat(wallet): 实现钱包余额充值功能和优化金额数据类型
- 将所有金额相关字段从Double改为BigDecimal以提高精度
- 新增充值订单的完整业务流程：创建、支付、查询、取消
- 在钱包余额服务中添加充值相关的接口定义和实现
- 更新订单控制器移除无用依赖并优化代码结构
- 扩展服务器资源实体和DTO支持续费周期配置
- 重构订单金额处理逻辑适配新的数据类型
- 添加钱包余额变更详情记录功能
- 实现推广记录中的金额精度优化

+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/dto/ServerResourcesRenewalDto.java
@@ -1,5 +1,6 @@
 package com.ruoyi.system.domain.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.AllArgsConstructor;
@@ -48,6 +49,6 @@ public class ServerResourcesRenewalDto {
     private Integer acceptablePriceIncreasePct;
 
     @ApiModelProperty("价格快照")
-    private Double currentPriceSnapshot;
+    private BigDecimal currentPriceSnapshot;
 
 }
